Building out Dosage information

+++ server/routes/comments.py
@@ -25,6 +25,9 @@ class CommentsByID(Resource):
         if not comment:
             return unfound_error('Comment')
         
+        if comment.user_id != session.get('user_id') and current_user.admin !="1":
+            return unauth_error
+        
         comment.comment = data['comment']
 
         db.session.commit()
@@ -37,6 +40,9 @@ class CommentsByID(Resource):
         
         if not comment:
             return unfound_error('Comment')
+        
+        if comment.user_id != session.get('user_id') and current_user.admin !="1":
+            return unauth_error
          
         db.session.delete(comment)
         db.session.commit()
@@ -50,12 +56,14 @@ class UserComments(Resource):
         
         if not user:
             return unfound_error('User')
+        
+        if user.id != session.get('user_id') and current_user.admin != "1":
+            return unauth_error
                 
         comments = [comment.to_dict() for comment in user.comments]
         return comments, 200
             
         
-    
 class UserCommentsByID(Resource):
     def get(self, id, comment_id):
         user = get_first(User, 'id', id)
@@ -64,6 +72,9 @@ class UserCommentsByID(Resource):
         
         if not user or not comment:
             return unfound_error('Item')
+        
+        if user.id != session.get('user_id') and current_user.admin != "1":
+            return unauth_error
                 
         if comment.user_id is not user.id:
             return {'error':'This comment was not posted by this user. '}, 409
@@ -80,6 +91,9 @@ class UserCommentsByID(Resource):
         
         if not user or not comment:
             return unfound_error('Item')
+        
+        if user.id != session.get('user_id') and current_user.admin != "1":
+            return unauth_error
                 
         if comment.user_id is not user.id:
             return {'error':'This comment was not posted by this user. '}, 409
@@ -97,6 +111,9 @@ class UserCommentsByID(Resource):
         
         if not user or not comment:
             return unfound_error('Item')
+        
+        if user.id != session.get('user_id') or current_user.admin != "1":
+            return unauth_error
         
         db.session.delete(comment)
         db.session.commit()
@@ -152,7 +169,10 @@ class RecipeCommentsByID(Resource):
             return unfound_error('Recipe')
             
         if comment.recipe_id is not id:
-            return unrelated_err('Comment', 'Recipe') 
+            return unrelated_err('Comment', 'Recipe')
+        
+        if comment.user_id != session.get('user_id') and current_user.admin != "1":
+            return unauth_error
         
         comment.comment = data['comment']
                         
@@ -170,6 +190,9 @@ class RecipeCommentsByID(Resource):
             
         if comment.recipe_id is not id:
                 return unrelated_err('Comment', 'Recipe') 
+        
+        if comment.user_id != session.get('user_id') and current_user.admin != "1":
+            return unauth_error
                 
         db.session.delete(comment)
         db.session.commit()
