+++ server/routes/herbs.py
@@ -1,6 +1,6 @@
 from flask_restful import Resource
 from config import api, db
-from flask import request, session, abort
+from flask import request, session
 from models.models import Herb, Property, Dosage
 from .helpers import get_current_user, get_all, unauth_error, unfound_error, get_first, deleted_msg
 
@@ -58,18 +58,19 @@ class HerbsByID(Resource):
 
         if not herb:
             return unfound_error('Herb')
-       
-        herb.name = data['name']
-        herb.latin_name = data['latin_name']
-        herb.description = data['description']
-        herb.warnings = data['warnings']
-        herb.image_url = data['image_url']
+        
+        if herb.entered_by_id != session.get('user_id') and current_user.admin != "1":
+            return unauth_error
+        
+        protected_attributes = ['id', 'entered_by_id', 'entered_on']
+
+        for key, value in data.items():
+            if key not in protected_attributes:
+                setattr(herb, key, value)
 
         if 'property_ids' in data:
             property_ids = data['property_ids']
-
             unique_property_ids = set(property_ids)
-
             properties = Property.query.filter(Property.id.in_(unique_property_ids)).all()
             herb.properties = properties
 
@@ -87,7 +88,10 @@ class HerbsByID(Resource):
         current_user = get_current_user()
 
         if not herb:
-            return unfound_error('Herb')  
+            return unfound_error('Herb')
+
+        if herb.entered_by_id != session.get('user_id') and current_user.admin != "1":
+            return unauth_error  
                 
         for ingredient in herb.ingredients:
             db.session.delete(ingredient)
@@ -99,6 +103,8 @@ class HerbsByID(Resource):
         db.session.commit()
             
         return deleted_msg('Herb')
-    
+
+
+
 api.add_resource(Herbs, '/api/herbs')
 api.add_resource(HerbsByID, '/api/herbs/<int:id>')
