+++ server/routes/recipes.py
@@ -54,9 +54,12 @@ class RecipesByID(Resource):
 
         if not recipe:
             return unfound_error('Recipe')
+        
+        if recipe.entered_by_id != session.get('user_id') and current_user.admin != "1":
+            return unauth_error 
             
-        recipe.name = data['name']
-        recipe.directions = data['directions']
+        for key, value in data.items():
+            setattr(recipe,key, value)
 
         if 'ingredients' in data:
             ingredients_data = data['ingredients']
@@ -87,6 +90,9 @@ class RecipesByID(Resource):
 
         if not recipe:
             return unfound_error('Recipe')
+        
+        if recipe.entered_by_id != session.get('user_id') and current_user.admin != "1":
+            return unauth_error 
         
         for ingredient in recipe.ingredients:
             db.session.delete(ingredient)
